+++ dev_notes/_run_check.R
@@ -1,6 +1,0 @@
-suppressPackageStartupMessages({
-  library(devtools)
-})
-res <- devtools::check(args = "--no-manual", error_on = "warning",
-                       quiet = FALSE)
-print(res)