+++ dev_notes/_run_visit_data_test.R
@@ -1,6 +1,0 @@
-suppressPackageStartupMessages({
-  devtools::load_all(quiet = TRUE)
-})
-res <- testthat::test_file("tests/testthat/test-visit-data.R",
-                           reporter = testthat::SummaryReporter$new())
-invisible(res)